requirements.txt
streamlit
keras
numpy
pillow
h5py